requirements.txt
python-telegram-bot
binance
ta
pandas
numpy
python-dotenv
joblib
requests